git-svn-id: http://springhead.info/spr2/Springhead2/trunk@2455 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Framework/FWScene.h
@@ -48,7 +48,7 @@ public:
 	virtual GRSceneIf*	GetGRScene(){ return grScene; }
 	virtual void		SetGRScene(GRSceneIf* s){ grScene = s; }
 	virtual FWObjectIf* CreateObject(const PHSolidDesc& soliddesc = PHSolidDesc(), const GRFrameDesc& framedesc = GRFrameDesc());
-	virtual int NObject()const{return fwObjects.size();}
+	virtual int NObject()const{return (int)fwObjects.size();}
 	virtual FWObjectIf** GetObjects(){return (FWObjectIf**)&*fwObjects.begin();}
 	virtual void Sync();
 	virtual void Step();
